Reload.txtが無い時の処理を追加

+++ Home.py
@@ -154,5 +154,6 @@ while True:
                 break
         display_window.close()
 
-os.remove('Reload.txt')
+if os.path.exists('Reload.txt'):
+    os.remove('Reload.txt')
 window.close()
